+++ setup.py
@@ -1,6 +1,22 @@
 #!/usr/bin/env python
 
+import os
+import torch
+from setuptools import setup, find_packages
+from torch.utils.cpp_extension import BuildExtension, CppExtension
+
+cmdclass = {}
+cmdclass['build_ext'] = BuildExtension
+
+
 import setuptools
+
+ext_modules = [
+    CppExtension(name='torch_blocksparse_cpp_utils',
+                 sources=['csrc/utils.cpp'],
+                 extra_compile_args={'cxx': ['-O2',
+                                             '-fopenmp']})
+]
 
 setuptools.setup(
     name             = 'torch-blocksparse',
@@ -10,9 +26,11 @@ setuptools.setup(
     maintainer       = 'Philippe Tillet',
     maintainer_email = 'ptillet@g.harvard.edu',
     install_requires = ['triton', 'torch'],
-    packages         = ['torch_blocksparse'],
     url              = 'https://github.com/ptillet/torch-blocksparse',
     test_suite       = 'nose.collector',
     tests_require    = ['nose', 'parameterized'],
-    license          = 'MIT'
+    license          = 'MIT',
+    packages         = find_packages(exclude=["csrc"]),
+    ext_modules      = ext_modules,
+    cmdclass         = cmdclass
 )
